requirements.txt
flask==3.1.2
PyGithub==2.8.1
requests==2.32.5
gunicorn==23.0.0
python-dotenv==1.1.1
